delete and update mutation added

+++ src/main/java/com/example/project/Person.java
@@ -16,7 +16,7 @@ public class Person {
 
     @Id
     @Column
-    private Long id;
+    private Integer id;
     @Column
     private String name;
 
@@ -35,7 +35,7 @@ public class Person {
         this.name = name;
     }
 
-    public Person(Long id, String name) {
+    public Person(Integer id, String name) {
 
         this.name = name;
     }
@@ -45,11 +45,11 @@ public class Person {
         return "Person [ name=" + name + "]";
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
